requirements.txt
streamlit
google-generativeai
supabase
pandas
numpy
extra-streamlit-components
pyarrow
openai
requests
python-dotenv
python-docx
pypdf
openpyxl
